cloud_functions/carga_inicial_calendar/requirements.txt
pandas
datetime
google-cloud-storage
fsspec
gcsfs
pandas-gbq